+++ assets/js/dashboard/filters.js
@@ -33,8 +33,16 @@ function filterText(key, value, query) {
   if (key === "browser") {
     return <span className="inline-block max-w-sm truncate">Browser: <b>{value}</b></span>
   }
+  if (key === "browser_version") {
+    const browserName = query.filters["browser"] ? query.filters["browser"] : 'Browser'
+    return <span className="inline-block max-w-sm truncate">{browserName}.Version: <b>{value}</b></span>
+  }
   if (key === "os") {
     return <span className="inline-block max-w-sm truncate">Operating System: <b>{value}</b></span>
+  }
+  if (key === "os_version") {
+    const osName = query.filters["os"] ? query.filters["os"] : 'OS'
+    return <span className="inline-block max-w-sm truncate">{osName}.Version: <b>{value}</b></span>
   }
   if (key === "country") {
     const allCountries = Datamap.prototype.worldTopo.objects.world.geometries;