+++ assets/js/dashboard/query.js
@@ -32,7 +32,9 @@ export function parseQuery(querystring, site) {
       'referrer': q.get('referrer'),
       'screen': q.get('screen'),
       'browser': q.get('browser'),
+      'browser_version': q.get('browser_version'),
       'os': q.get('os'),
+      'os_version': q.get('os_version'),
       'country': q.get('country'),
       'page': q.get('page')
     }